🤖 tests(streaming): drop EventEmitter-era scaffolding, share sink helper
